Move utils into tasks/data.py
